# bff/api-docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backend For Frontend API",
    "description": "API documentation for the BFF service.",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
    "/actors": {
      "get": {
        "tags": [
          "Actors"
        ],
        "operationId": "get_all_actors",
        "responses": {
          "200": {
            "description": "List all actors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WrapperActor"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Actors"
        ],
        "operationId": "create_actor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WrapperNewActor"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Create a new actor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrapperActor"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/actors/{actor_id}": {
      "get": {
        "tags": [
          "Actors"
        ],
        "operationId": "get_actor_by_id",
        "parameters": [
          {
            "name": "actor_id",
            "in": "path",
            "description": "ID of the actor to retrieve",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get actor by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrapperActor"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Actors"
        ],
        "operationId": "update_actor",
        "parameters": [
          {
            "name": "actor_id",
            "in": "path",
            "description": "ID of the actor to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WrapperNewActor"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Update actor by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrapperActor"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Actors"
        ],
        "operationId": "delete_actor",
        "parameters": [
          {
            "name": "actor_id",
            "in": "path",
            "description": "ID of the actor to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete actor by ID",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "minimum": 0
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/movies": {
      "post": {
        "tags": [
          "Movies"
        ],
        "operationId": "create_movie",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WrapperNewMovie"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Movie created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrapperMovie"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/movies/{movie_id}": {
      "get": {
        "tags": [
          "Movies"
        ],
        "operationId": "get_movie_by_id",
        "parameters": [
          {
            "name": "movie_id",
            "in": "path",
            "description": "ID of the movie to retrieve",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get movie by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrapperMovie"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Movies"
        ],
        "operationId": "update_movie",
        "parameters": [
          {
            "name": "movie_id",
            "in": "path",
            "description": "ID of the movie to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WrapperNewMovie"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Update movie by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrapperMovie"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Movies"
        ],
        "operationId": "delete_movie",
        "parameters": [
          {
            "name": "movie_id",
            "in": "path",
            "description": "ID of the movie to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete movie by ID",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "minimum": 0
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "WrapperActor": {
        "type": "object",
        "required": [
          "created_at",
          "first_name",
          "id",
          "last_name"
        ],
        "properties": {
          "bio": {
            "type": [
              "string",
              "null"
            ]
          },
          "birth_date": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "last_name": {
            "type": "string"
          }
        }
      },
      "WrapperMovie": {
        "type": "object",
        "required": [
          "created_at",
          "id",
          "slug",
          "title",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "duration_minutes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "mpaa_rating": {
            "type": [
              "string",
              "null"
            ]
          },
          "release_date": {
            "type": [
              "string",
              "null"
            ]
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "WrapperNewActor": {
        "type": "object",
        "required": [
          "first_name",
          "last_name"
        ],
        "properties": {
          "bio": {
            "type": [
              "string",
              "null"
            ]
          },
          "birth_date": {
            "type": [
              "string",
              "null"
            ]
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          }
        }
      },
      "WrapperNewMovie": {
        "type": "object",
        "required": [
          "slug",
          "title"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "duration_minutes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "mpaa_rating": {
            "type": [
              "string",
              "null"
            ]
          },
          "release_date": {
            "type": [
              "string",
              "null"
            ]
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
    }
  }
}